chore: migrate OperatorForOperationOut schema to v2 and clean up operator.py
